conversion warning fix

+++ src/mtsmsme/processor/PRNTSM.cpp
@@ -73,7 +73,7 @@ void PRNTSM::CONTINUE_received(uint8_t cdlen,
     cont.setComponent(0,iid);
     std::vector<unsigned char> rsp;
     tco->encoder.encode_resp(cont,rsp);
-    tco->SCCPsend(raddrlen,&raddr[0],laddrlen,laddr,rsp.size(),&rsp[0]);
+    tco->SCCPsend(raddrlen,&raddr[0],laddrlen,laddr,(uint16_t)rsp.size(),&rsp[0]);
 
     smsc_log_debug(logger,
                    "tsm.PRNTSM otid=%s receive CONTINUE with component, INVOKE_RES sent",
@@ -107,8 +107,8 @@ void PRNTSM::TEndReq()
                    laddrlen,dump(laddrlen,laddr).c_str());
   smsc_log_error(logger,
                  "PRN[%d]={%s}",
-                 data.size(),dump(data.size(),&data[0]).c_str());
-  tco->SCCPsend(raddrlen,raddr,laddrlen,laddr,data.size(),&data[0]);
+                 data.size(),dump((uint16_t)data.size(),&data[0]).c_str());
+  tco->SCCPsend(raddrlen,raddr,laddrlen,laddr,(uint16_t)data.size(),&data[0]);
 }
 void PRNTSM::TBeginReq(uint8_t  cdlen, uint8_t* cd, uint8_t  cllen, uint8_t* cl)
 {
